[refactor] Allow other packages depend on Config without depending on bot
There is a FR upcoming which will use Telegram API (not Telegram Bot
API) which will rely on the config and the chat structure
